requirements.txt
click==7.1.2
GitPython==3.1.11
PyYAML==5.3.1
pytest==6.2.1
pytest-mock==3.4.0
